+++ gnl/get_next_line_utils.c
@@ -12,9 +12,9 @@
 
 #include "get_next_line.h"
 
-size_t	ft_strlen(char *str)
+int	ft_strlen(char *str)
 {
-	size_t	c;
+	int	c;
 
 	c = 0;
 	if (!str)
